+++ nautilus_trader/adapters/interactive_brokers/execution.py
@@ -829,6 +829,12 @@ class InteractiveBrokersExecutionClient(LiveExecutionClient):
                 if order.totalQuantity == UNSET_DECIMAL
                 else Quantity.from_str(str(order.totalQuantity))
             )
+
+            if total_qty <= 0.0:
+                # This can be caused by a partially filled entry bracket order and SL triggered.
+                self._log.warning(f"IB order with totalQuantity <= 0, skipping: {order.__dict__}")
+                return
+
             price_magnifier = self.instrument_provider.get_price_magnifier(
                 nautilus_order.instrument_id,
             )